requirements.txt
Flask>=3.0.0
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-core>=0.2.0
langchain-community>=0.2.0
python-dotenv>=1.0.0
faiss-cpu>=1.8.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0
requests>=2.32.0
gunicorn>=21.0.0